backend/requirements.txt
flask==3.1.3
httpx==0.28.1
python-dotenv==1.2.2
